Move KegBase to separate file
